requirements.txt
PyGithub
jinja2